Fix bugs and Add separate routes for server js's readability
