+++ server.js
@@ -8,12 +8,6 @@ const session = require('express-session');
 const Mongostore = require('connect-mongo');
 const mongoose = require('mongoose');
 const cors = require('cors')
-const OpenAI = require("openai");
-const openai = new OpenAI(
-    { apiKey: process.env.OPENAI_API_KEY }
-);
-const constants = require('./constants.json');
-
 
 /* const saltRounds = */
 const port = process.env.PORT || 3000;
@@ -32,18 +26,15 @@ const node_session_secret = process.env.NODE_SESSION_SECRET;
 app.set('views', path.join(__dirname, 'views'));
 app.set("view engine", "ejs");
 
-var { database } = include('databaseConnection');
 
-const userCollection = database.db(mongodb_database).collection('users');
-
-mongoose.connect(mongodb_uri)
-	.then(() => {
-		console.log('Connected to MongoDB')
-	})
-	.catch((error) => {
-		console.error('Error connecting to MongoDB:', error)
-	}
-)
+mongoose.connect(mongodb_uri, { useNewUrlParser: true, useUnifiedTopology: true })
+    .then(() => {
+        console.log('Connected to MongoDB')
+    })
+    .catch((error) => {
+        console.error('Error connecting to MongoDB:', error)
+    }
+    )
 
 var mongoStore = Mongostore.create({
     mongoUrl: mongodb_uri,
@@ -79,134 +70,24 @@ isValidSession = (req) => {
 
 /* ------- all routes ------- */
 
-const authRouter = require ("./routes/authentication")
+const authRouter = require("./routes/authentication")
+const aiAdvisorRouter = require("./routes/aiAdvisor")
 
 app.use("/", authRouter);
+app.use("/", aiAdvisorRouter);
+
 
 app.get('/home', (req, res) => {
     res.render('main');
-
 });
 
 app.get('/addFriend', (req, res) => {
     res.render('addFriend');
 });
+
 app.get('/addGroup', (req, res) => {
     res.render('addGroup');
 });
-app.get('/AI', (req, res) => {
-    res.render('aiAdvisor');
-});
-
-
-app.post('/advisor', async function (req, res) {
-    console.log(req.body);
-    let { userMessages, assistantMessages } = req.body
-
-    let messages = [
-        { role: "system", content: constants.SYSTEM_COMMENT },
-        { role: "user", content: constants.USER_COMMENT },
-        { role: "assistant", content: constants.ASSISTANT_COMMENT },
-    ]
-
-    while (userMessages.length != 0 || assistantMessages.length != 0) {
-        if (userMessages.length != 0) {
-            messages.push(
-                JSON.parse('{"role": "user", "content": "' +
-                    String(userMessages.shift()).replace(/\n/g, "") + '"}')
-            )
-        }
-        if (assistantMessages.length != 0) {
-            messages.push(
-                JSON.parse('{"role": "assistant", "content": "' +
-                    String(assistantMessages.shift()).replace(/\n/g, "") + '"}')
-            )
-        }
-    }
-
-    app.post('/advisor', async function (req, res) {
-        let { userMessages, assistantMessages } = req.body
-    
-        let messages = [
-            { role: "system", content: constants.SYSTEM_COMMENT },
-            { role: "user", content: constants.USER_COMMENT },
-            { role: "assistant", content: constants.ASSISTANT_COMMENT },
-        ]
-    
-        while (userMessages.length != 0 || assistantMessages.length != 0) {
-            if (userMessages.length != 0) {
-                messages.push(
-                    JSON.parse('{"role": "user", "content": "' +
-                        String(userMessages.shift()).replace(/\n/g, "") + '"}')
-                )
-            }
-            if (assistantMessages.length != 0) {
-                messages.push(
-                    JSON.parse('{"role": "assistant", "content": "' +
-                        String(assistantMessages.shift()).replace(/\n/g, "") + '"}')
-                )
-            }
-        }
-    
-        const maxRetries = 3;
-        let retries = 0;
-        let completion
-        while (retries < maxRetries) {
-            try {
-                completion = await openai.chat.completions.create({
-                    model: "gpt-4o",
-                    messages: messages
-                });
-                break;
-            } catch (error) {
-                retries++;
-                console.log(error);
-                console.log(`Error fetching data, retrying (${retries}/${maxRetries})...`);
-            }
-        }
-    
-        let chatGPTResult = completion.choices[0].message.content
-        question = messages.at(-1).content
-        answer = chatGPTResult
-        console.log(`Question: ${messages.at(-1).content}\nAnswer: ${chatGPTResult}`);
-    
-    
-        // await chatHistoryCollection.insertOne({ question, answer, timestamp: new Date() });
-    
-        // await sessionCollection.updateOne(
-        //     { sessionId: req.sessionID },
-        //     {
-        //         $push: { qaPair: { question: question, answer: chatGPTResult, timestamp: new Date() } },
-        //         $setOnInsert: { sessionId: req.sessionID, createdAt: new Date() }
-        //     },
-        //     { upsert: true }
-        // );
-    
-        res.json({ "assistant": chatGPTResult });
-    });
-    
-    const maxRetries = 3;
-    let retries = 0;
-    let completion
-    while (retries < maxRetries) {
-        try {
-            completion = await openai.chat.completions.create({
-                model: "gpt-4o",
-                messages: messages
-            });
-            break;
-        } catch (error) {
-            retries++;
-            console.log(error);
-            console.log(`Error fetching data, retrying (${retries}/${maxRetries})...`);
-        }
-    }
-
-    let chatGPTResult = completion.choices[0].message.content
-    console.log(chatGPTResult);
-    res.json({ "assistant": chatGPTResult });
-});
-
 
 // all unrelated routes
 app.get('*', (req, res) => {
